Remove support for directories configuration
Existing packages in Node.js/npm ecosystem use a "directories" property
to indicate that modules are in the "lib" directory, but Node.js does
not read this configuration, so the user still has to include "lib" in
module identifiers.  This is not according to spec, but the behavior of
ignoring the "directories" property has become a de-facto spec and these
packages will not load with Mr if we persist in reading
`directories.lib`.  Also, there's no point in fighting the
`node_modules` dependency packages directory configuration any longer.
I have lost that fight.