Fix loading of minified code
This is a partial soultion to the more general problem of knowing
whether a module should be loaded from a `.js` file with the same
identifier, or loaded from text and offerred up to middleware to provide
alternate compilers, translators, and optimizers.

In the future, there will be configuration information in `package.json`
that will provide better insight.  For now, we can just assume that
module identifiers that lack an extension, and the `.min` extension
should be followed by `.js`.

This also simplifies paths and extensions middleware into location
middleware.

These changes have been applied separately to the omnibus
configurability branch.