Add support for optimizers
Optimizers only run in production and can perform transformations in
text from modules of one type to another.  This feature will hopefully
enable a significant refactoring in Mop.

Also, restores ability to load JSON modules, with a spec to prevent
future regressions.

There is also some superficial restoration of the order in which
functions are declared, and reduction of duplication of the script
injection code.