Disable Firebug workaround that breaks modern Firefox
The workaround is not necessary in modern Firefox, and the argument name
interfered with conditional loading of Underscore in Emmet. I’ve fixed
the issue with the shadowed underscore variable name, and then commented
the workaround out entirely, leaving it in the source just in case it
needs to be uncommented for debugging under older Firebugs, and as a
historical note of how things were.