Don't load devDeps into mappings
When loaded as library in the browser (and perhaps in Node.js too) Mr
loads its own dependencies into the config mappings. This works for Q,
however this also loads the devDependencies which are generally not
installed when Mr is used, and shouldn't override the client's
dependencies anyway.

This commit puts Mr into "production" mode, which means it will not
read the devDependencies.